some more int vs. u-int fixes

+++ FDTD/operator.cpp
@@ -731,18 +731,18 @@ bool Operator::CalcEFieldExcitation()
 	vector<unsigned int> vDelay;
 	vector<unsigned int> vDir;
 	unsigned int ipos;
-	int pos[3];
+	unsigned int pos[3];
 	double coord[3];
 	double delta[3];
 	double amp=0;
 
-	for (pos[2]=0;pos[2]<(int)numLines[2];++pos[2])
+	for (pos[2]=0;pos[2]<numLines[2];++pos[2])
 	{
 		delta[2]=fabs(MainOp->GetIndexDelta(2,pos[2]));
-		for (pos[1]=0;pos[1]<(int)numLines[1];++pos[1])
+		for (pos[1]=0;pos[1]<numLines[1];++pos[1])
 		{
 			delta[1]=fabs(MainOp->GetIndexDelta(1,pos[1]));
-			for (pos[0]=0;pos[0]<(int)numLines[0];++pos[0])
+			for (pos[0]=0;pos[0]<numLines[0];++pos[0])
 			{
 				delta[0]=fabs(MainOp->GetIndexDelta(0,pos[0]));
 				coord[0] = discLines[0][pos[0]];
@@ -759,7 +759,7 @@ bool Operator::CalcEFieldExcitation()
 						CSPropElectrode* elec = prop->ToElectrode();
 						if (elec!=NULL)
 						{
-							if ((elec->GetActiveDir(n)) && (pos[n]<(int)numLines[n]-1))
+							if ((elec->GetActiveDir(n)) && (pos[n]<numLines[n]-1))
 							{
 								amp = elec->GetWeightedExcitation(n,coord)*delta[n]*gridDelta;
 								if (amp!=0)
@@ -824,7 +824,7 @@ bool Operator::CalcEFieldExcitation()
 //						cerr << n << " " << coord[0] << " " << coord[1] << " " << coord[2] << endl;
 						if (elec!=NULL)
 						{
-							if ((elec->GetActiveDir(n)) && (pos[n]<(int)numLines[n]-1))
+							if ((elec->GetActiveDir(n)) && (pos[n]<numLines[n]-1))
 							{
 								amp = elec->GetWeightedExcitation(n,coord)*deltaN*gridDelta;
 								if (amp!=0)
